requirements.txt
# ========================================
# DECAI ORACLE V2 - Python Dependencies
# ========================================
# Updated: 09/02/2026
# ========================================

# Web3 & Blockchain
web3>=6.0.0
eth-account==0.11.0
eth-utils>=2.3.0
# hexbytes version managed by eth-account

# Environment & Configuration
python-dotenv==1.0.0

# Machine Learning & Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
matplotlib==3.8.2
seaborn==0.13.1

# API & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.0.0
pydantic-settings

# Async HTTP (CRITICAL for V2 multi-source)
aiohttp>=3.9.0
pycoingecko==3.1.0

# Database & Cache
sqlalchemy>=2.0.0
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.39.2

# IPFS (Phase 2)
ipfshttpclient==0.8.0a2

# Security & Encryption
cryptography>=42.0.0

# Utilities
requests==2.31.0
python-dateutil>=2.8.0
celery==5.3.4

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint>=3.0.0

# Development
ipython>=8.0.0

# ========================================
# INSTALLATION INSTRUCTIONS
# ========================================
# 
# 1. Activate virtual environment:
#    Windows: .\venv\Scripts\activate
#    Linux/Mac: source venv/bin/activate
#
# 2. Install dependencies:
#    pip install -r requirements.txt
#
# 3. Verify installation:
#    python -c "import web3, aiohttp; print('✅ Dependencies OK')"
#
# ========================================
